frameworks/google-adk/requirements.txt
google-adk>=1.0.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-processor-baggage>=0.61b0
pytest>=7.0.0
pytest-asyncio>=0.23.0
